pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "canvas-dl"
version = "0.1.0"
description = "Canvas Files Downloader CLI for UVA"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "UVA Canvas Downloader" }
]
requires-python = ">=3.9"
dependencies = [
  "typer>=0.9.0",
  "rich>=13.0.0",
  "httpx>=0.27.0",
  "tenacity>=8.2.2",
  "pypdf>=4.2.0",
  "python-dotenv>=1.0.0",
  "platformdirs>=4.0.0",
  "questionary>=2.0.1",
  "tomli>=2.0.1; python_version < '3.11'",
  "tomli-w>=1.0.0; python_version < '3.11'"
]

[project.urls]
Homepage = "https://example.com"

[project.scripts]
canvas-dl = "canvas_dl.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.9"
strict = false
